Move type declaration to the nearest component file
