+++ client/src/lib/types.ts
@@ -1,7 +1,7 @@
 export interface CodeReview {
+    approvers?: string[];
     createdAt: string;
     id: number;
-    approvers?: string[];
     jiraTicket?: string;
     jiraTicketLinkUrl?: string;
     note?: string;
@@ -17,26 +17,4 @@ export interface CodeReview {
 
 export interface User {
     displayName: string;
-}
-
-export interface ExpandedRowProps {
-    data: CodeReview;
-    onUpdate: CallableFunction;
-    user: User;
-}
-
-export interface DataViewProps extends ExpandedRowProps {
-    onError: CallableFunction;
-}
-
-export type DataEditProps = DataViewProps;
-
-export interface UpdateEventProps {
-    data: CodeReview;
-    action: string;
-}
-
-export interface ChannelFilterProps {
-    onSelected: CallableFunction;
-    selectedChannel: string;
 }
